extract out server function into it's own file
